Fixed type in method name. Added setState
setState method will allow the app to tell the Domain when the state changes, and allows the domain to perform internal actions accordingly.
In the future the states could have constructors to force correct info to be provided when changing state. (i.e. replaying asks for the initial world conditions)