Removed unneeded flag, added final tick flag
The purpose of undone flag was to not undo the same tick twice, but this is being checked elsewhere, making the flag useless.

Program needs some way to know if the tick it received is the last one. This is communicated through this finalTick flag, but it needs to be maintained by some other component because domain has no way of knowing which update will be its last.