Updated teleporter to comply with new terrain interface requirement
This is a unique case where the terrain that the game object ends up is actually not the terrain the world moves it too.
It isn't even necessarily the terrain at the teleporter's link location, in the case of a teleporter that sends you straight into another teleporter.

This code can cope with that