Implemented Next Level functionality, added in sounds and other item interactions
Made it so that one can move from the test level to level 1 upon entering the exit tile. There is one bug where an exception is thrown upon exiting, needs further investigation.

Added in sounds when picking up items and various other interactions such as pushing the block.

Also made it so that the inventory is printed to console upon pickup or consumption of item.

Note that you will need to add a dummy file path to the persistency's Concrete Persister since the code doesn't load in files properly.