Implement timer, inventory, number of collected treasures, help pages, keyboard actions for the GUI, and made changes to App to work with new Recorder implementation
Implemented display treasure and inventory display, timer, pause dialog, keyboard actions, log messaging system for the Fuzzer, and implemented pre/post conditions to ensure healthy state upon transition.

Number of treasure and the player's inventory is now reported in a visible text area.

Timer updates every tick and compensates for the delay to ensure that it always times out after 60 or so seconds.

Pause dialog is shown in place of the timer label when a pause occurs.

Keyboard actions as according to the handout as well as additional ones were implemented. The KeyListener was replaced with an ActionMap system which can utilise key binds.
The user has the ability to toggle pause with the keyboard as well as the button.

A log messaging system for the Fuzzer, which logs all warnings and notifications into a fixed size array which has a LogMessage added to it everytime some notification occurs.
LogMessages contain a warning flag to show that it is warning message.

Many pre condition checks, post condition state changes were added to ensure robust state transitions.

Various bug fixes such as adding a fixed size onto the action queue so that the fuzz doesn't overflow it, adding in the ability to click anywhere on the frame to regain focus, clearing the inventory upon player losing, and resetting the game to the same level upon level lose.

Created a proxy class to give instances of to recorder and changed the GameLoop and corresponding actions to match this.

Changed the formula used to calculate the auto-playback speed to match up with the new Recorder implementation.